+++ web/sw.js
@@ -20,6 +20,7 @@ const SHELL = [
   './icons.js',
   './i18n.js',
   './presets.js',
+  './whatsnew.js',
   './manifest.webmanifest',
   './icons/icon-192.png',
   './icons/icon-512.png',